+++ src/main.rs
@@ -1,9 +1,15 @@
+mod juno_manager;
 mod simulate_transaction;
+mod trace_block;
+mod trace_transaction;
 
+use simulate_transaction::simulate_main;
+use trace_block::block_main;
+use trace_transaction::transaction_hash_main;
 
 #[tokio::main]
 async fn main() {
-    simulate_transaction::transaction_hash_main().await.unwrap();
-    simulate_transaction::block_main().await.unwrap();
-    simulate_transaction::simulate_main().await.unwrap();
+    transaction_hash_main().await.unwrap();
+    block_main().await.unwrap();
+    simulate_main().await.unwrap();
 }
